Single Folder zip file or create powershell zip script
Use this to create a powershell script to run and regularly back up a
folder to the same backup location. Created for backing up school class
materials regularily without having to select and zip the folders every
time. Just let the script run and create a backup zip of the folder
heirarchy